debugged 1-binary.c

+++ 0x1E-search_algorithms/1-binary.c
@@ -13,19 +13,19 @@
  */
 int binary_search(int *array, size_t size, int value)
 {
-	int left = 0, right = size - 1, mid;
+	size_t left = 0, right = size - 1, mid;
 
 	if (array && size > 0)
 	{
 		while (left <= right)
 		{
-			int index;
+			size_t index;
 
 			mid = (right + left) / 2;
 			printf("Searching in array: ");
 			for (index = left; index < right; index++)
 				printf("%d, ", array[index]);
-			printf("%d\n", right);
+			printf("%ld\n", right);
 			if (array[mid] > value)
 				right = mid - 1;
 			else if (array[mid] < value)
